+++ http4k-core/src/main/kotlin/org/http4k/core/Http4k.kt
@@ -5,10 +5,11 @@ import org.http4k.routing.RoutingHttpHandler
 typealias HttpHandler = (Request) -> Response
 
 fun interface Filter : (HttpHandler) -> HttpHandler {
-    companion object
+    companion object {
+        @JvmStatic
+        fun create(fn: (HttpHandler) -> HttpHandler): Filter = Filter { next -> fn(next) }
+    }
 }
-
-fun Filter.Companion.create(fn: (HttpHandler) -> HttpHandler): Filter = Filter { next -> fn(next) }
 
 val Filter.Companion.NoOp: Filter get() = Filter { next -> { next(it) } }