+++ http4k-security-oauth/src/main/kotlin/org/http4k/security/oauth/server/refreshtoken/RefreshTokens.kt
@@ -9,13 +9,11 @@ import org.http4k.security.oauth.server.TokenRequest
 import org.http4k.security.oauth.server.UnsupportedGrantType
 import org.http4k.security.oauth.server.accesstoken.GrantType
 
-interface RefreshTokens {
+fun interface RefreshTokens {
     fun refreshAccessToken(clientId: ClientId, tokenRequest: TokenRequest, refreshToken: RefreshToken): Result<AccessToken, AccessTokenError>
 
     companion object {
-        val unsupported = object : RefreshTokens {
-            override fun refreshAccessToken(clientId: ClientId, tokenRequest: TokenRequest, refreshToken: RefreshToken): Result<AccessToken, AccessTokenError> = Failure(UnsupportedGrantType(GrantType.RefreshToken.rfcValue))
-        }
+        val unsupported = RefreshTokens { _, _, _ -> Failure(UnsupportedGrantType(GrantType.RefreshToken.rfcValue)) }
     }
 }