+++ http4k-security-oauth/src/main/kotlin/org/http4k/security/openid/RequestJwts.kt
@@ -3,7 +3,7 @@ package org.http4k.security.openid
 import org.http4k.security.State
 import org.http4k.security.oauth.server.AuthRequest
 
-interface RequestJwts {
+fun interface RequestJwts {
     fun create(authRequest: AuthRequest, state: State, nonce: Nonce?): RequestJwtContainer
 }