+++ CHANGELOG.md
@@ -4,7 +4,8 @@ This list is not intended to be all-encompassing - it will document major and br
 changes with their rationale when appropriate:
 
 ### v3.262.0 (uncut)
-- **http4k-*** : Upgrade some dependency versions.
+- **http4k-*** : Upgrade some dependency versions, including Kotlin to 1.4.10
+- **http4k-*** : Take advantage of Kotlin Functional Interfaces, including for Filter.
 - **http4k-testing-kotest - Possible Break (on Jackosn)**: Remove a `haveBody` matcher which uses `Matcher<JsonNode>` directly, because of a bug in Kotest:
 https://github.com/kotest/kotest/issues/1727